refactored viewer.config.ts to split type definition into separate file
